// src/app/model/user.model.ts
export class User {

    id: number = 0;
    email: string = "";
    password: string = "";
    confirmPassword: string = "";
    fullName: string = "";
    phoneNumber: string = "";
    accessToken: string = "";
    about: string = "";
    imageFile: string = "";
    firstName: string = "";
    lastName: string = "";
    company: string = "";
    gender: string = "";
    userId: string = "";
    loginCount: 0;
    blogCount: number = 0;
}